requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
reportlab
fpdf
opencv-python-headless
